MINOR: mux-quic: simplify decode_qcs API
Slightly modify decode_qcs function used by transcoders. The MUX now
gives a buffer instance on which each transcoder is free to work on it.
At the return of the function, the MUX removes consume data from its own
buffer.

This reduces the number of invocation to qcs_consume at the end of a
full demuxing process. The API is also cleaner with the transcoders not
responsible of calling it with the risk of having the input buffer
freed if empty.

(cherry picked from commit 62eef85961f4a2a241e0b24ef540cc91f156b842)
[wt: ctx updates due to some parts that were already integrated in
     qcc_decode_qcs(), h3_headers_to_htx(), and h3_decode_qcs()]
Signed-off-by: Willy Tarreau <w@1wt.eu>