MEDIUM: fd: add the tgid to the fd and pass it to fd_insert()
The file descriptors will need to know the thread group ID in addition
to the mask. This extends fd_insert() to take the tgid, and will store
it into the FD.

In the FD, the tgid is stored as a combination of tgid on the lower 16
bits and a refcount on the higher 16 bits. This allows to know when it's
really possible to trust the tgid and the running mask. If a refcount is
higher than 1 it indeed indicates another thread else might be in the
process of updating these values.

Since a closed FD must necessarily have a zero refcount, a test was
added to fd_insert() to make sure that it is the case.

(cherry picked from commit 9464bb1f05b5e0046716b4573a567d3450ac7604)
[wt: this is necessary to fix a design bug introduced in 2.5. The patch
 was simplified as 2.6 does not yet support thread groups, only group 1
 is added and the fd_insert() API does not change]
Signed-off-by: Willy Tarreau <w@1wt.eu>