MINOR: fd: add fd_get_running() to atomically return the running mask
The running mask is only valid if the tgid is the expected one. This
function takes a reference on the tgid before reading the running mask,
so that both are checked at once. It returns either the mask or zero if
the tgid differs, thus providing a simple way for a caller to check if
it still holds the FD.

(cherry picked from commit ceffd17f52b6b1aa481365fc6f9b88e8efc436e8)
[wt: this is necessary to fix a design bug introduced in 2.5]
Signed-off-by: Willy Tarreau <w@1wt.eu>