MINOR: fd: make fd_clr_running() return the previous value instead
It's an AND so it destroys information and due to this there's a call
place where we have to perform two reads to know the previous value
then to change it. With a fetch-and-and instead, in a single operation
we can know if the bit was previously present, which is more efficient.

(cherry picked from commit d6e198761281e755e5b96272ed2a751df096efdf)
[wt: this is necessary to fix a design bug introduced in 2.5; ctx
 adjustment for tid_bit instead of ti->ltid_bit]
Signed-off-by: Willy Tarreau <w@1wt.eu>