BUG/MINOR: quic: Wrong status returned by qc_pkt_decrypt()
This bug came with this big commit:
     "MEDIUM: quic: xprt traces rework"

This is the <ret> variable value which must be returned by most of the xprt functions.
This leaded packets which could not be decrypted to be parsed, with weird frames
to be parsed as found by Tristan in GH #1808.

To be backported where the commit above was backported.

(cherry picked from commit 11a6f4007b908b49ecd3abd5cd10fba177f07c11)
Signed-off-by: Willy Tarreau <w@1wt.eu>